Changes to be committed: 	modified:   index.php 	modified:   public/css.panel/general.css 	renamed:    public/img.users/1.png -> public/img.users/1.jpg 	renamed:    public/img.users/2.png -> public/img.users/2.jpg 	new file:   public/img.users/bpohfo6963wq5ywmaf1yah12y.png 	new file:   public/img/doctor1_respaldo.jpg 	new file:   public/img/doctor2_respaldo.jpg 	modified:   public/js.general/general.js 	new file:   public/js.panel/clientes.js 	new file:   public/js.panel/horas.js 	deleted:    public/js.panel/links.js 	new file:   public/js.panel/mensajes.js 	deleted:    public/js.panel/plans.js 	new file:   public/js.panel/servicios.js 	modified:   public/js.panel/slider.js 	new file:   public/js.panel/social.js 	deleted:    public/js.panel/table.js 	modified:   public/js.panel/user.js 	deleted:    public/js.panel/users.js 	modified:   src/assets/db.sql 	new file:   src/dao/ClienteDao.php 	new file:   src/dao/MensajeDao.php 	modified:   src/routes/panel.php 	modified:   src/routes/services.php 	new file:   src/services/cliente.service.php 	new file:   src/services/mensaje.service.php 	modified:   src/services/servicio.service.php 	modified:   src/services/slider.service.php 	modified:   src/services/user.service.php 	modified:   src/templates/panel.component/head.php 	modified:   src/templates/panel.component/header.php 	modified:   src/templates/panel.component/sidebar.php 	renamed:    src/templates/panel.pages/plans.php -> src/templates/panel.pages/clientes.php 	modified:   src/templates/panel.pages/home.php 	renamed:    src/templates/panel.pages/links.php -> src/templates/panel.pages/horas.php 	modified:   src/templates/panel.pages/info.php 	new file:   src/templates/panel.pages/mensajes.php 	renamed:    src/templates/panel.pages/users.php -> src/templates/panel.pages/servicios.php 	modified:   src/templates/panel.pages/slider.php 	new file:   src/templates/panel.pages/social.php 	modified:   src/templates/panel.pages/template.php 	new file:   src/templates/panel.pages/user.php

+++ src/services/servicio.service.php
@@ -79,8 +79,18 @@ class ServicioService
             $servicio_imagen = $current_servicio['servicio_imagen'];
 
             if (isset($_FILES['servicio_imagen'])) {
-                deleteFile('./public/img.servicios/' . $servicio_imagen);
-                $servicio_imagen = uploadFIle($_FILES['servicio_imagen'], './public/img.servicios/');
+                if ($_FILES['servicio_imagen']['tmp_name'] != "" or $_FILES['servicio_imagen']['tmp_name'] != null) {
+                    if (
+                        $servicio_imagen != '1.png' &&
+                        $servicio_imagen != '2.png' &&
+                        $servicio_imagen != '3.png' &&
+                        $servicio_imagen != '4.png' &&
+                        $servicio_imagen != ''
+                    ) {
+                        deleteFile('./public/img.servicios/' . $servicio_imagen);
+                    }
+                    $servicio_imagen = uploadFIle($_FILES['servicio_imagen'], './public/img.servicios/');
+                }
             }
 
             $servicio = $servicioDao->update(
@@ -122,7 +132,15 @@ class ServicioService
                 echo json_encode($result);
                 exit();
             }
-            deleteFile('./public/img.servicios/' . $servicio['servicio_imagen']);
+            if (
+                $servicio['servicio_imagen'] != '1.png' &&
+                $servicio['servicio_imagen'] != '2.png' &&
+                $servicio['servicio_imagen'] != '3.png' &&
+                $servicio['servicio_imagen'] != '4.png' &&
+                $servicio['servicio_imagen'] != ''
+            ) {
+                deleteFile('./public/img.servicios/' . $servicio['servicio_imagen']);
+            }
             $servicio = $servicioDao->deleteById($servicio_id);
             $result['status'] = 'success';
             $result['message'] = 'Servicio eliminado correctamente';
